+++ services/get-stripe.ts
@@ -4,11 +4,9 @@
 import { Stripe, loadStripe } from '@stripe/stripe-js'
 
 let stripePromise: Promise<Stripe | null>
-const getStripe = () => {
+export default function getStripe() {
   if (!stripePromise) {
     stripePromise = loadStripe(process.env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY!)
   }
   return stripePromise
 }
-
-export default getStripe
